Split slack-ping into a wrapper, service class and constants
no ref

The single-file service read module-level singletons directly, so its
unit test had to stub globals and fish listeners back out of the events
stub to invoke them.

Follows the common service wrapper convention: index.js wires the
dependencies inside init() and constructs a SlackPingService class that
receives everything through its constructor. The test now builds the
class with injected fakes and calls the handlers directly. Listener
registration is normalised to the stable-bound-listener pattern used by
indexnow-ping; behaviour is otherwise unchanged and boot.js already
calls init().
