+++ ghost/core/core/server/services/slack-ping/constants.js
@@ -1,0 +1,19 @@
+const messages = {
+    requestFailedError: 'The {service} service was unable to send a ping request, your site will continue to function.',
+    requestFailedHelp: 'If you get this error repeatedly, please seek help on {url}.'
+};
+
+const defaultPostSlugs = [
+    'welcome',
+    'the-editor',
+    'using-tags',
+    'managing-users',
+    'private-sites',
+    'advanced-markdown',
+    'themes'
+];
+
+module.exports = {
+    messages,
+    defaultPostSlugs
+};
